effects: the operation vocabulary is not node's
`OpResult`, `IoError`, `IoErrorInfo`, `IoChannel`, `IoResult` and the
`ioError` / `toIoError` / `isNotFound` constructors were declared in
`effects/node`, but nothing in them names a host: "the runner cannot
dispatch this" and "the host tried and failed" are the two ways any
operation goes wrong, on any host. They move to `effects/` beside
`NotImplemented`, which `OpResult` is defined in terms of.

The misfiling already had a victim: `effects/memory/types.ts` — which
has no host at all — imported `OpResult` from `../node/types.ts`. It
now takes it from the layer it belongs to.

`effects/node` re-exports every moved name, so the several dozen
modules that reach for them through it are unchanged, and an
operation's declaration still reads as one vocabulary. Groundwork for
step 4 of emergent_testing/todo/share-browser-console-runner.md: a
second host's operations cannot be typed while the types they are
written in live in the first host's module.
