+++ fjs/effects/types.ts
@@ -21,11 +21,11 @@ import type {
  * `error(notImplemented(command))` through the command's own output — so an
  * operation whose return admitted no error would be a hole in that mechanism:
  * there would be nowhere to put the refusal. Every *host* operation already
- * returned `OpResult<…>` or `IoResult<…>` when this constraint was added, and
- * so did four of the six declared inside proofs — through a bare `Result`
- * rather than either alias, since those two are node conveniences. The two in
- * `./proof.f.mjs` returned a bare `number`, and the commit that added the rule
- * rewrote them.
+ * returned {@link OpResult} or {@link IoResult} when this constraint was added,
+ * and so did four of the six declared inside proofs — through a bare `Result`
+ * rather than either alias, which were then declared in `./node/types.ts` and
+ * so read as node conveniences. The two in `./proof.f.mjs` returned a bare
+ * `number`, and the commit that added the rule rewrote them.
  *
  * It is also the latch the whole representation now rests on. An operation
  * *cannot* be declared infallible, so every {@link Effect} built from one has a
@@ -64,6 +64,71 @@ export type Operation =
  * one type — see {@link Effect}.
  */
 export type NotImplemented = readonly['notImplemented', string]
+
+/**
+ * The result of an operation with no failures of its own: it either produces
+ * its value or reports that the runner does not implement it.
+ *
+ * Every operation's return type is a `Result`, including the ones that cannot
+ * fail on their own terms — an operation left on a raw contract would be a hole
+ * in the error channel, and a runner may omit a handler for any of them.
+ */
+export type OpResult<T> = Result<T, NotImplemented>
+
+/**
+ * A host failure, normalized: whatever the runtime threw reduced to a
+ * serializable record. `code` is the OS error code when the host supplied one
+ * (`'ENOENT'`, `'EEXIST'`), absent otherwise.
+ *
+ * It is a tagged tuple for the same reason {@link NotImplemented} is — the two
+ * share an error channel, and the tag is what tells them apart. That
+ * distinction is the whole reason this type exists: with a bare `unknown`
+ * error, `NotImplemented | unknown` collapses to `unknown` and a program can no
+ * longer tell "this runner cannot do it" from "the host tried and failed".
+ *
+ * Normalizing also keeps the channel serializable. A thrown `Error` carries a
+ * stack, a `cause`, and arbitrary own properties; none of it survives a wire
+ * hop, and a runner in another process could not reproduce it.
+ */
+export type IoError = readonly['ioError', IoErrorInfo]
+
+export type IoErrorInfo = {
+    readonly code?: string
+    readonly message: string
+}
+
+/**
+ * The error channel of anything that performs host IO: a normalized host
+ * failure, or the report that the runner does not implement the operation.
+ *
+ * It is one name rather than a union spelled at each site, and that is a
+ * migration property rather than brevity. An effect that does no IO *yet* is
+ * one added `readFile` away from doing some, and if each signature names its
+ * own errors, that one change walks up every enclosing signature — the failure
+ * mode that sank `throws` clauses elsewhere, where engineers eventually
+ * declared everything throwing rather than maintain the cascade. Declaring the
+ * standard channel once is that concession made deliberately: an IO-touching
+ * effect says it fails *the way host IO fails*, and gaining a new way to do so
+ * changes nothing above it.
+ *
+ * It is not a licence to widen. An operation with failures of its own extends
+ * the channel (`IoChannel | ParseError`), and a computation whose errors are
+ * genuinely narrower should say so — this is the default for IO, not a ceiling.
+ *
+ * **It is not node's, though it was declared there.** Nothing in either half
+ * names a host: a runner that cannot dispatch and a host that tried and failed
+ * are the two ways any operation goes wrong, on any host. Living in
+ * `./node/types.ts` meant that `./memory/types.ts` — which has no host at
+ * all — reached into the node module for {@link OpResult}, and that a second
+ * host's operations could not be typed without doing the same.
+ */
+export type IoChannel = NotImplemented | IoError
+
+/**
+ * The result of an operation that performs host IO: its value, a normalized
+ * host failure, or the missing-handler report.
+ */
+export type IoResult<T> = Result<T, IoChannel>
 
 /**
  * An `Effect<O, T, E>` is the raw value: a {@link Pure} thunk yielding
